+++ src/Client/ButtonModule.cpp
@@ -1,0 +1,12 @@
+#include "ButtonModule.hpp"
+#include "ButtonModuleNode.hpp"
+
+ModuleNode* ButtonModule::getNode()
+{
+    return ButtonModuleNode::create(this);
+}
+
+bool ButtonModule::shouldSave()
+{
+    return false;
+}