+++ src/GUI/BetterMouseDispatcher.cpp
@@ -1,5 +1,7 @@
 #include "BetterMouseDispatcher.hpp"
 
+#ifndef GEODE_IS_IOS
+
 bool BetterMouseDispatcher::dispatchScrollMSG(float x, float y)
 {
     auto pRet = CCMouseDispatcher::dispatchScrollMSG(x, y);
@@ -11,3 +13,5 @@ bool BetterMouseDispatcher::dispatchScrollMSG(float x, float y)
 
     return pRet;
 }
+
+#endif