+++ src/GUI/BetterMouseDispatcher.hpp
@@ -10,7 +10,11 @@ struct MouseDispatcher
     static inline std::vector<CCMouseDelegate*> betterMouseDispatcherDelegates = {};
 };
 
+#ifndef GEODE_IS_IOS
+
 class $modify (BetterMouseDispatcher, CCMouseDispatcher)
 {
     bool dispatchScrollMSG(float x, float y);
 };
+
+#endif